Initial support for Rust on Zephyr
This is based off of the initial support merged into the
[collab-rust](https://github.com/zephyrproject-rtos/zephyr/tree/collab-rust)
branch on the main Zephyr branch, which includes two PRs, one adding
initial support and the other adding bindgen.

Currently the docs have been moved into the README.rst in the module,
and the same is placed here, although twister doesn't currently run
samples that are in modules.

Signed-off-by: David Brown <david.brown@linaro.org>