Bring in the Apache-2.0 license
All of the code developed for the Zephyr on Rust will be under the
Apache-2.0 license, as this work is part of the Zephyr project.

Some dependencies linked into the application may have relaxations on
these license requirements, as Apache-2.0 or MIT is a common license in
the Rust community.

Signed-off-by: David Brown <david.brown@linaro.org>