zephyr: Make symbols from portable-atomic available
Provide the atomic types  from portable-atomic in
`zephyr::sync::atomic`, and, if allocation is enabled, provide
`zephyr::sync::Arc`.  The `alloc::arc` will only be available on Zephyr
targets that have atomic intrinsics, and the portable one can be
supported on any Zephyr target.  There are some limitations described in
the documentation.

Signed-off-by: David Brown <david.brown@linaro.org>