zephyr: Add alignment helper
The Rust attribute that indicates structure alignment only supports
alignment values that are numeric constants.  This makes it difficult to
set alignment based on a value coming from the build environment, which
is commonly done.

Implement a bit of a workaround with an `AlignAs` type that is
parameterized with the alignment.  Parameters can be compile-time
constants not just numeric constants.  This works by having instances
for each of the desired alignments that each have the numeric constant
as the alignment.  The instances can be expanded if necessary.

Signed-off-by: David Brown <david.brown@linaro.org>