Upgrade crate versions to match Zephyr
Upgrade the versions of all crates in this module to match the current
Zephyr version of 3.7.0.  It isn't completely clear if this is the right
thing to do, as Zephyr doesn't maintain semantic versioning.  But having
the numbers match will help make it clear which version of Zephyr goes
with which version of this crate.

Signed-off-by: David Brown <david.brown@linaro.org>