zephyr: object: Add kboj_define support for arrays of stacks
Create a const constructor function for building the stack arrays.  This
requires unsafe, and can't use MaybeUninit, because the array version of
this is still unsable.  Just use zero initialization, and make sure the
loop initializes everything.

Use this function in the kobj_define macro in order to allow apps to
define arrays of thread stacks, in addition to arrays of threads.

Signed-off-by: David Brown <david.brown@linaro.org>