zephyr: sys: sync: Mutex: Implement Send/Sync
Although it is tradition in Rust to have types such as Semaphores and
Mutexes not implement Sync, these primitives, built as thing wrappers
become fairly useless.  Sharing them would require `Arc`, which requires
alloc. Presumably someone wanting to use lower level primitives would
also not likely be wanting to use allocation.

For the most part, these primtives have as their real purpose to be used
in the implementation of the higher level synchronization primtives,
such as sync::Mutex.

unlock is unsafe because it is required to only call unlock on the same
thread that locked.

Signed-off-by: David Brown <david.brown@linaro.org>