zephyr: Remove 'unsafe' from sys::Mutex::unlock
Although this function has constraints on how it can be used (the thread
that calls unlock must also have called lock).  However, according to
the documentation, it detects this, and returns an error.  As such, the
wrapper in Rust does not need to be `unsafe` but can merely reflect that
error code in the `Result` that it returns.

Signed-off-by: David Brown <david.brown@linaro.org>