zephyr: Make gpio interface much more "unsafe"
Gpios in Zephyr are inherently unsafe.  There is a shared controller
that is not just used by the pins defined here, but extensively across
various drivers.

As such, make all of the gpio pin operations themselves unsafe.

We can help, a little bit, to at least enforce uniqueness with the
Rust drivers that use gpios by requiring them to take a mutable instance
of `GpioToken`, which has a singleton getter.

Signed-off-by: David Brown <david.brown@linaro.org>