zephyr-build: Include DT node presence configs
Generate a full list of nodes that are present in the given device tree,
and provide a tool that build.rs in the app can use to make these
active.  This will allow conditionals like:

    #[cfg(dt = "aliases::led")]

to be used, which will make it possible to handle nodes being present or
not in the DTS.

See a subsequent patch to the blinky sample for an example of usage.

Signed-off-by: David Brown <david.brown@linaro.org>