CI: Clean up the work the CI build does
Add -M full to the twister build so that it cleans up space as it goes.
This should allow the build to run to completion.

Add -j4 to the build. Although the CI images currently have 4 cpus, this
will help protect us against future ones that have more.

Give twister an explicit list of platforms.  This fixes two issues.  One
is that the output is getting flooded with messages about skipped tests.
The other is this allows us to add platforms that qemu seems to be
ignoring entirely.

Signed-off-by: David Brown <david.brown@linaro.org>