platforms: Remove mps2
The gpio device tree entries for the mps2 are defined with a
`#gpio-cells` value of 1, despite these values not being interpreted by
the driver, but by the devicetree code.  I'm not sure if these actually
work with any of the demos, as it is unclear what the macros would do
with this.  It doesn't give us a value to use for dt_flags.

If someone wants to put in some effort to fix this, feel free.  But I
don't think the problem is on the Rust side.

Signed-off-by: David Brown <david.brown@linaro.org>