zephyr: device: uart: Add IRQ mode
Add an `.into_irq()` to the Uart device that turns it into an interrupt
driven interface.  Currently, read is implemented, with a `.try_read()`
method that will try, with a timeout, to read data from the interface.

The methods are all marked as 'unsafe' currently, until a more thorough
safety analysis can be made.

Signed-off-by: David Brown <david.brown@linaro.org>