zephyr: device: uart: Fix lifetimes of 'inner' method on UartIrq
This method had a lifetime error, and allowed the inner uart to be
leaked beyond its lifetime.

This should prevent the returned device from outliving the UartIrq, or
even being used after mutable methods on the UartIrq have been used.
For now, don't even return a mutable method until we evaluate what
should be permissible.  This at least allows reading the line status.

In addition, since the return Uart is not owned by the caller, it is not
permissible to call any of the `into_*` methods.

Signed-off-by: David Brown <david.brown@linaro.org>