zephyr: device: uart: irq: Replace write with async API
Update the writing to be an async interface.  Instead of a blocking
write, there is a `write_enqueue` to add the message, and a `write_wait`
which is a blocking wait for a write to handle.

For this to work, the UartIrq needs to own the buffer for the duration
of the write.  At this point, the buffers are just `Vec<u8>`, but a
future change will make this more general, so that, for example, Pooled
buffers can be used.  The buffer should never be dropped within this
code.

Signed-off-by: David Brown <david.brown@linaro.org>