zephyr: Logging through Zephyr log
Implement a simplistic mechanism that is able to log through Zephyr's
logging mechanism.

This change does not attempt to demangle the #define nightmare that is
Zephyr's logging, but just formats to a string, and sends that directly
through a C helper function.

As such, this performs at least one allocation/free, possibly with some
reallocs, depending on how accurate the allocation estimate is, and
whether adding a null terminator requires an additional allocation.

Logs are also all recorded under a single registered log, in the rust
main file.

Signed-off-by: David Brown <david.brown@linaro.org>