zephyr: sync: channel: Wrap channel implementation with "flavor"
Taken from crossbeam channel, instead of directly storing the queue into
the Receiver and Sender, instead wrap this in a enum, which will allow
us to have different kinds of implementations.

This is done via enum rather than traits as it would otherwise have to
be dynamic, and a small dynamic decision is generally faster than the
vtable that would result from using `dyn`.

Signed-off-by: David Brown <david.brown@linaro.org>