cmake: Add rustdoc target
Add a target to the build `rustdoc` that will generate the rust
documentation.  The documentation will be placed in a target-specific
directory under `build/rust/target/...`.

There is some duplication in the cmake file, as I can't figure out how
to factor this out into a variable and maintain the proper quoting of
the space separated strings.

Signed-off-by: David Brown <david.brown@linaro.org>