zephyr: sys: Create Timer
Add a wrapper `Timer<T>` to wrap Zephyr's `k_timer` type.  This is
typed, so that we can constrain the type used for callbacks.  If
callbacks are not used, it is sufficient to use this as `Timer<()>`.

Currently, this timer type can only be dynamically allocated, as the
static allocation and the typed interact weirdly.  This can be solved by
exposing the inner timer wrapper, and having the user instantiate the
Timer with a `new_from` method, similar to how this is done with Mutex.

It might make sense to have the sys wrapper be exposed, and a more
strongly typed version under `sync`.  However, the difference is very
minimal, and would mainly serve to make initialization slightly easier
for the static case.

Signed-off-by: David Brown <david.brown@linaro.org>