zephyr: Create Timer
Add a set of wrappers for Zephyr's timer.  Timers start as
`StoppedTimer`, which can be either dynamic or static (through object).
This can then be started either with `start_simple` or `start_callback`
which will return a `SimpleTimer` or a `CallbackTimer`.  This can then
be turned back into stopped timers with a `stop` method.

Signed-off-by: David Brown <david.brown@linaro.org>