Move to semantic versioning
An earlier change updated all of the crate version numbers to match a
Zephyr release.  This is problematic for several reasons:

- Zephyr releases don't correlate with semantic versions.  In general,
  it is probably best to assume every Zephyr release is a semantic
  change, and should have a major version bump.
- Cargo does make some assumptions about the meaning of semantic
  versions (for example, that it can use a version that would be
  considered semantically equivalent.
- The Rust support consists of multiple crates, and there is no
  particular reason to tie the version numbers of those together.

At this point, the Rust support for Zephyr is experimental, and
definitely needs a 0.x.y version number.  Revert everything back to
0.1.0.  As this stabilizes more, we can start to come up with
appropriate numbers.

Although we are quite a w ways from it making sense to include any of
this in the crate ecosystem, it will still be helpful to have meaningful
numbers if we ever do decide to do that.

Signed-off-by: David Brown <david.brown@linaro.org>