cmake: Consistent expansion of include path
Accept both space and semicolon separated paths and defines from cmake.
The add_custom_command really wants to replace the semicolons with
spaces, whereas the 'file' command doesn't.  To make this work, have the
build.rs split on either space or semicolon.

This fixes a problem where `cargo check` and `rust-analyzer` fail to run
correctly when the Zephyr build is configured for release, and they are
trying to do a debug build.  With this fix, the debug build from cargo
will work, allowing the analysis or verification.

Signed-off-by: David Brown <david.brown@linaro.org>