zephyr: work: Support showing work in Segger SystemView
If `CONFIG_SEGGER_SYSVIEW` is enabled, add a few calls to the workqueue
scheduling to record these as task switches.  There is still a problem
with the task names showing up, but this does generally show scheduled
work as separate tasks in System View.

The main issues where this lacks is that if the Zephyr scheduler
schedules something else, and then returns to the work queue, that will
be shown as being done on the work queue thread instead of the
particular work.

Signed-off-by: David Brown <david.brown@linaro.org>