zephyr: work: Create ContextExt
Because Zephyr's workqueue re-scheduling is based on the Context struct,
not the waker, we need a way to get back to our `WorkInfo` struct that
this Context is embedded in.

Create a `ContextExt` that adds some methods to the `Context` to be able
to indicate when the scheduler should reschedule this work.

Signed-off-by: David Brown <david.brown@linaro.org>