samples: bench: Benchmark a "base semaphore"
This is a comparison benchmark of implementing deferred initialization
of Zephyr kernel objects from Rust.  The idea here is that instead of
the current `Fixed` enum, which contains either a reference to a static,
or a Pin<Box<T>> of the object, we pair the objects with an atomic
pointer variable indicating the pointer state.  It starts as a null
pointer, and will result in the object being initialized when it is
first used.  The pointer is then replaced with the address of `Self` so
that later code can detect an attempt to more the object.

Signed-off-by: David Brown <david.brown@linaro.org>