github: workflow: docs: Push 'main' docs to gh-pages
Distinguish between pull requests, and pushes to 'main'.  The pull
request docs are published on s3, with a mechanism to retire those after
a time.

Pushes to 'main' (happens when code merges) replace the documentation on
gh-pages.  These are available for anyone wanting to see "the latest
docs".

Signed-off-by: David Brown <david.brown@linaro.org>

docs: Add latest commit to top of index

Insert `git show --stat HEAD` into the template for the top-level docs,
to make it easy to see what a build was generated for.

Signed-off-by: David Brown <david.brown@linaro.org>