docs: Update top-level index
Add a bit of additional information to the top level page.  Most uses of
the docs will link directly to relevant pages, but the top level page
will be seen by anyone navigating directly to the doc pages.

This also contains the text of the latest commit, to help anchor what
these docs are for.

Signed-off-by: David Brown <david.brown@linaro.org>