zephyr: object: Implement new ZephyrObject wrapper
Implement a new wrapping system for Zephyr Objects.  Instead of trying
to either use a static or a pinned-box, use an atomic to detect
initialization, as well as improper moves.

This allows for objects to have `const` constructors, allowing them to
be declared as simple statics and just shared.

Subsequent patches will implement this for various primitives.

Signed-off-by: David Brown <david.brown@linaro.org>