zephyr: sys: sync: Make Semaphore constructors `const`
Move Semaphores to the new `ZephyrObject` system, allowing the
constructor to be const.  Unfortunately, `Result` isn't generally useful
with const, and as such, this changes the API for `Semaphore::new()` to
just return the semaphore instead of a Result.  Instead, if a semaphore
constructor has inconsitent arguments, it will panic.  Given that
semaphore constructors are generally quite simple, this is a minor
issue.

However, it will involve removing a lot of `.unwrap()` or `.expect(...)`
calls on semaphores.

The samples have been changed for the API change, but mostly not changed
to take advantage of the easier ways to use them.  It is generally no
longer necessary to put semaphores in an Arc or Rc, but just to allocate
them in something like a StaticCell, and then pass them around by
reference.

Signed-off-by: David Brown <david.brown@linaro.org>