samples: bench: Make Semaphores static
Although there is still a lot of allocation left (the crossbeam API for
channels is kind of built around it), convert the Semaphores in the
benchmark to be static.  This reduces overhead a little bit, because the
reference counts of Arc don't need to be maintained, and also
demonstrates how StaticCell can be used to allocate things like arrays
of objects statically.

Signed-off-by: David Brown <david.brown@linaro.org>