zephyr: Implement `executor-zephyr` to use with Embassy
This implements an executor for Embassy that runs a set of async tasks
on a single Zephyr thread.  The executor will suspend the thread when
there is no work to do, resuming when there is work.

It is permissible to run multiple executors, on different threads, to
support a hybrid async cooperative and priority-based scheduling.

Signed-off-by: David Brown <david.brown@linaro.org>