samples: embassy: Update Embassy Demo for Zephyr executor
Update the embassy demo to use the Zephyr executor.  This performs a
ping-ping test using two executors, with one end of the responder
running on another thread.

On the rp2040, the round trip time with `executor-threaded` is about
12us, with `executor-zephyr`, all on a single thread, it is about 15us,
and with pairs of task running across differrent threads, the time is
about 26us.

Signed-off-by: David Brown <david.brown@linaro.org>