tests: drivers: gpio-async: Test the async gpio operations
This is a basic test to test for the waiting gpios.  Currently, this
test requires a specific hardware setup.  To test on the rpi-pico, it is
necessary to connect GPIO-0 to GPIO-6 with a button or switch.
Running should then print Pressed/Released on a debounced version of
that switch, with the code not polling the gpio while waiting.

Signed-off-by: David Brown <david.brown@linaro.org>