zephyr: device: Add a possible static element with each device
Allow the dt-rust.yaml file to specify a `static_type` value for device
nodes.  The generated code for the nodes will reserve a static value of
that type, and pass it into the constructor.  There is a NoStatic which
is the default.  As it is zero size, it will not cause any code to be
generated for the types that do not need static.

Fix the flash device to accept this argument, but the fix for gpio will
come in a subsequent patch.

Signed-off-by: David Brown <david.brown@linaro.org>