zephyr: mark old async methods as deprecated
This work-queue based async executor, although functional, has some
significant problems. Most notably, it doesn't use wakers, and as such,
is incompatible with other utilities that have been written to be used
with Async.  Trying to combine them will, in fact, result in undefined
behavior.

Now that the embassy-based executor-zephyr is in main, mark these as
deprecated so that code can gradually move to newer solutions.

Signed-off-by: David Brown <david.brown@linaro.org>