zephyr: Add async 'wait_for_high' and 'wait_for_low' to gpio
Implement a general async `wait_for_high` and `wait_for_low` to the
GpioPin interface.  The waiting is supported at the level of the gpio
driver. Each Gpio instance has an array of `AtomiWaker` to hold the
wakers that are waiting on the pins, as well as the callback handler
from using the pins in interrupt mode in Zephyr.

Although most of the examples of gpios in Zephyr use irq-gpios instead
of gpios in the custom properties, there isn't anything that happens
when this is done, and regular 'gpios' properties work just fine with
interrupts.

Signed-off-by: David Brown <david.brown@linaro.org>