zephyr: Remove work-queue executor
The work-queue-based executor was an attempt to implement a Rust
executor that worked with Zephyr work queues.  Although it was possible
to make something that kind of worked, and did have the advantage of
being able to use a few Zephyr primitives, it suffers from several
problems:

- It isn't very efficient, basically combining the overhead of async
  scheduling, triggered work, and the Zephyr thread scheduler.
- It plays poorly with any other use of async.  Since it is useful to be
  able to use external crates for async utilities, this normally just
  results in undefined-behavior.

Now that we have another executor (built from Embassy), this code is
also mostly unneeded.

Signed-off-by: David Brown <david.brown@linaro.org>