zephyr: embassy: Decouple embassy-time and Zephyr time
Don't require that the two time bases be the same.  This allows
applications to work using the default embassy-time base of 1Mhz.  There
is a performance cost to the conversion (which depends on the exact
ratios).  If the time bases are the same (which would be common for an
application built for a single target), then no conversion is needed.

Signed-off-by: David Brown <david.brown@linaro.org>