zephyr: net: add initial support for UDP sockets
Initial implementation of a Rust wrapper around the Zephyr networking
API for UDP sockets, loosely based on the UdpSocket in the Rust std
library. It's possible to create/bind a socket and sendto/recvfrom data,
but not much else.

There is also some supporting functionality for converting IP addresses
and socket addresses between the C representation and Rust's core::net
types.

Signed-off-by: Matt Rodgers <mrodgers@witekio.com>