Add rust-toolchain file
This specifies the exact tools that need to be present in order to build
these projects.  This should affect all of the samples.

Currently, we require Rust 1.85 for a few reasons.  There have been some
subtle changes to warnings that cause build failures with older version.
Specifically, the 2024 edition requires certain attributes to explicitly
be marked as unsafe, but the construct to mark them as unsafe is only
present in the 2024 compiler.  There are also some subtle changes to
warnings that affect our builds.

Signed-off-by: David Brown <david.brown@linaro.org>