cmake: Fix dependency problem
For normal builds, the offsets.h file gets generated before the Rust
target builds.  But, when trying to build just the docs, the dependency
is not explicit, and the file is just not generated.

Fix this by making the dependency explicit.  Like the others, we can't
use the top-level define name, and have to state the dependency target
by expanded name.  This is due to the "fascinating" scoping rules in
cmake.

Signed-off-by: David Brown <david.brown@linaro.org>