zephyr: Specify minimum rust version
To allow the 2024 edition of Rust to be used, one of our macros needs a
new syntax, introduced in Rust 1.85 to declare that it is known that the
macro expansion is unsafe.

This could work without the minimum rust version, as long as the macro
isn't used, but is needed for tests, and Rust stable has already moved
to 1.86, so it isn't an onerous requirement.

Signed-off-by: David Brown <david.brown@linaro.org>