zephyr: work: Generalize Work pointer type
First step in generalizing the worker type.  Place the inner pointer
type into a trait, and allow the work methods to work on pointers of
that type.

This continues to work the same as before with `Pin<Arc<Work<T>>>`.  It
needs a bit more massaging to be able to work with static pointers,
namely that the constructor has to be able to be static, and as such,
the initialization can't happen until it has been pinned down.  This
will likely change the "new" API.

Signed-off-by: David Brown <david.brown@linaro.org>