work: Specialize Arc-based Work
Instead of hard-coding `Pin<Arc<Work<T>>>`, have a constructor that
wraps the `Arc<Work<T>>` in a new type `ArcWork`.  This has the methods
for submitting work on this particular pointer type.

This leaves room for `StaticWork` which will have a static reference.

Removes all references to Pin.  Now that the Work queue itself is stored
in a ZephyrObject, we have a mechanism to panic on move, which although
a runtime check, does catch issues where values have been moved.

Signed-off-by: David Brown <david.brown@linaro.org>