zephyr-sys: Handle deflected GPIO constants
bindgen is willing to perform some simple math when expanding macros,
but, does not see through the BIT macro.

As a workaround, until we can come up with a better solution, export
`ZR_` prefixed versions that are consts so that they will get exported.
Unfortunately, this changes the visible exported name (which should be
fixed as well)

Signed-off-by: David Brown <david.brown@linaro.org>