zephyr: work: Allow struct to have a additional fields
Structs in the C world assume zero initialization, and therefore adding
fields is not considered an API change.  Support this for an upcoming
API change in the work queue config, by deriving default for these
structs, and using this zero-init as a default.  This should support
fields being added to this struct. The rust code will zero init the
unknown fields, which will have the same behavior as the C code.

Signed-off-by: David Brown <david.brown@linaro.org>