zephyr: gpio: Simplify usage
The gpio driver's 'token' to ensure uniqueness doesn't really help with
correctness (as C code can use the gpio device as well).  Short of
re-defining the gpio interfaces, remove this token entirely, and make
only the `get_instance` of gpio unsafe, to indicate that the safety of
GPIO operations is dependent on the underlying controller.

This should improve the hygeine of the GPIO API significantly, and allow
uses, such as waiting on multiple pins, to be possible.

Signed-off-by: David Brown <david.brown@linaro.org>