dt-rust: Fix "Root" node detection
The labels node should occur only at the root node in the Rust-converted
devicetree.  However, when translating the format of the yaml file to
use explicit types, the rule for this node didn't end up in an array.
It isn't clear why this doesn't generate an error, but making it an
explicit array does fix the issue.

Without this fix, the devicetree.rs that is generated has a duplicate
copy of the labels module under every node.  This fixes it to only occur
at the top level.

Fixes: #113

Signed-off-by: David Brown <david.brown@linaro.org>