samples/tests: Capture Cargo.lock
A recent short-term broken release of a transient dependency broke CI
builds of Zephyr with Rust support. Although it was fixed shortly
thereafter, it is a problem we'd like to avoid.

Until now, we've avoided committing the lock files.  This change commits
the lock files for each sample and test, building them with versions
that have been tested.

Signed-off-by: David Brown <david.brown@linaro.org>